Move files around to make more sense in the structure
