+++ spec/components/routes/join-game.spec.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import JoinGame from '../../../src/components/routes/join-game.jsx';
+import JoinGame from '../../../src/routes/join-game.jsx';
 
 describe('<JoinGame />', () => {
   test('renders', () => {
