+++ src/components/rules.jsx
