+++ src/routes/home.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import {Link} from 'react-router-dom';
 
-import GameManager from '../../lib/game-manager';
+import GameManager from '../lib/game-manager';
 
-import Button from '../shared/button.jsx';
-import Rules from './components/rules.jsx';
+import Button from '../components/shared/button.jsx';
+import Rules from '../components/rules.jsx';
 
 const Home = () => (
   <div>
