+++ src/routes/new-game.jsx
